requirements.txt
# Core ML and Computer Vision
ultralytics>=8.0.0
opencv-python>=4.8.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=10.0.0

# Data Processing and Analysis
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.10.0

# Video Processing
imageio>=2.31.0
imageio-ffmpeg>=0.4.8

# Tracking and Analytics
filterpy>=1.4.5
scikit-learn>=1.3.0

# Visualization and UI
streamlit>=1.28.0
plotly>=5.17.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
requests>=2.31.0
